moved the spectral filtering method to utils, to keep feature detection clean
